requirements.txt
streamlit
pandas
matplotlib
wordcloud
numpy
openai
google-api-python-client
google-auth
google-cloud-storage
